Integration of Erant's Wii DVD-Loading via libdi and Wii Zip ROM support (Use: 7za a -tzip -mx=9 rom.zip rom.z64). GC ROM-Cache needs a re-work to add in Zip support, then some of the nastyness of my quick-fix hacks to get this compiling for both GC and Wii will be gone!

+++ trunk/main/gc_dvd.h
@@ -5,6 +5,7 @@
 #ifndef GC_DVD_H
 #define GC_DVD_H
 
+#include <stdint.h>
 //Used in ISO9660 Parsing
 #define NO_FILES -1
 #define NO_ISO9660_DISC -2
@@ -13,11 +14,11 @@
 
 void dvd_motor_off();
 unsigned int dvd_get_error(void);
-int dvd_read_directoryentries(unsigned int offset, int size);
+int dvd_read_directoryentries(uint64_t offset, int size);
 void read_directory(int sector, int len);
-int read_safe(void* dst, int offset, int len);
+int read_safe(void* dst, uint64_t offset, int len);
 int read_direntry(unsigned char* direntry);
-int read_sector(void* buffer, int sector);
+int read_sector(void* buffer, uint32_t sector);
 int dvd_read(void* dst,unsigned int len, unsigned int offset);
 extern unsigned char sector_buffer[2048] __attribute__((aligned(32)));
 int is_unicode;
